integrate Phyto and rework Zoo + more verbose
Phyto and Zoo are basically the same data format,
so both can be processed in the the same way.
Zoo needed some reworking to follow the same
pipeline, and the Phyto dataset brought up
a need to generalize a few functions
such as `clean_sp_names` and `surv_to_wide`.