Fix bizarre bug: sdmpredictors broke targets
When I called sdmpredictors, the function
worked in an interactive session.

However, when I used it in a pipeline,
the target calling it would be saved as
a NULL object.

The fix was to supress warnings on the
sdmpredictors call.

I have no idea why this happened, or
why the fix works. My best guess is that
the warnings in sdmpredictors were not
normal messages, and caused R to
drop out early and return NULL.

I won't report it, it seems like a niche
problem, and I can't provide any useful
reproducible code.