Allow workers to cd to the workdir
Also set runtime as a template option.

Because workers are submitted from
singularity via ssh, they always start from
the home directory.

Because the work dir is inside a temp directory
we need a way to pass in the work dir. I have
added a template option to pass and set the
worker directory.