Figured out bottleneck
about 10000 rows is large enough that loop overheads is not
the bottleneck. Above this point limiting factor
is GPU processing the matrix batches. Larger batches take
longer for the GPU, in roughly linear fashion.
nvitop suggests GPU is only using about 40% of processing
power.

Here we limit the size of batches to make lots of memory available
for parallel workers to use the GPU together.