Rework clustering methods
Add default variants of each clustering technique.

The optimal parameter search methods are renamed to ~.*optimal$~
The new default parameter methods are named ~.*default$~

For caster, the default is the mean similarity.
Motivated by Hubert Gamma using the mean similarity as the threshold
for increasing the score when sites are clustered together.

For apcluster, the default is q=0.5, or median similarity.
Motivated by the package authors suggesting it as a reasonable default
for more rather than fewer clusters. Also, the other reasonable
default suggested is q=0, but the minimum similarity is always 0
except in extreme cases, and probably not as informative.