Don't filter by minimum CoV
Implemented to drop species that do not respond to the environment by
being relatively constant everywhere.

However, species that do not respond to the environment are dropped by
GF anyway, R^2 is a measure of how well the variance was reduced
relative to the orginial variance in the data. There is no need  to hide low
variance species from GF.