Tweak rsync args to transfer faster
-I: don't skip based on time or size. Faster than -c, because it
doesn't have to read the whole file to generate a checksum
The cache is a single large file that certainly changed, always try to
transfer it

-i: output a list of files transferred
--inplace: update the target file by modification, rather than
creating a new file then deleting the old file.
Risky if multiple processes try to do the same thing in parallel, but
I don't expect that to be happening here.